% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foo_bar.R
\name{foo_bar}
\alias{foo_bar}
\title{foo_bar}
\usage{
foo_bar()
}
\description{
foo_bar
}
\details{
asdfasdflajsdfals dflasj dfaljsdf asldfj asdlfja sdfljas dflja sdfljas dflajs dflajsdf alsjf aljsdf
}
\examples{
foo_bar()
}
