% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foobar3-package.R
\docType{package}
\name{foobar3-package}
\alias{foobar3-package}
\alias{foobar3}
\title{foobar3 package}
\description{
asdlfjasdlfa fljas falsjfd dsalfj asljf asljf asljdf alfkj asldfj asldfj aslfj aslfjaslfasdf slj lj afs
}
\author{
Scott Chamberlain \email{myrmecocystus@gmail.com}
}
\keyword{package}
